FPVLT-493: Moved setTaskCompletion method to CaseUtils
